bugs squashed - cart icon animation added - cart item counter woking
