+++ src/css/style.css
@@ -1,66 +1,82 @@
 :root {
+  /* Font settings */
   --font-body: Arial, Helvetica, sans-serif;
   --font-headline: Haettenschweiler, "Arial Narrow Bold", sans-serif;
-  /* colors */
+
+  /* Color palette */
   --primary-color: #f0a868;
   --secondary-color: #525b0f;
   --tertiary-color: #a4b8c4;
   --light-grey: #d0d0d0;
   --dark-grey: #303030;
 
-  /* sizes */
+  /* Font sizes */
   --font-base: 18px;
   --small-font: 0.8em;
   --large-font: 1.2em;
 }
-* {
-  box-sizing: border-box;
-}
-body {
-  margin: 0;
-  font-family: var(--font-body);
-  font-size: var(--font-base);
-  color: var(--dark-grey);
-}
-img {
-  max-width: 100%;
-}
+
+      /* Reset and global styles */
+      * {
+        box-sizing: border-box;
+      }
+
+      body {
+        margin: 0;
+        font-family: var(--font-body);
+        font-size: var(--font-base);
+        color: var(--dark-grey);
+      }
+
+      img {
+        max-width: 100%;
+      }
+
+/* Header styling */
 header {
   display: flex;
   justify-content: space-between;
   padding: 0 10px;
 }
-.logo {
-  line-height: 60px;
-  width: 245px;
-  overflow: hidden;
-  display: flex;
-  font-size: 30px;
-  text-align: right;
-  font-family: var(--font-headline);
-}
-.logo img {
-  width: 60px;
-  height: 60px;
-}
-.logo a {
-  text-decoration: none;
-  color: var(--font-body);
-}
+
+      /* Logo styling */
+      .logo {
+        line-height: 60px;
+        width: 245px;
+        overflow: hidden;
+        display: flex;
+        font-size: 30px;
+        text-align: right;
+        font-family: var(--font-headline);
+      }
+
+      .logo img {
+        width: 60px;
+        height: 60px;
+      }
+
+      .logo a {
+        text-decoration: none;
+        color: var(--font-body);
+      }
+
+/* Hero section */
 .hero {
   position: relative;
 }
+
 .hero .logo {
   position: absolute;
   left: calc(50% - 60px);
   bottom: -10px;
   background-color: white;
 }
+
+/* Square logo variant */
 .logo--square {
   font-size: var(--small-font);
   width: 110px;
   height: 80px;
-
   flex-direction: column;
   align-items: flex-end;
   box-shadow: 1px 1px 4px var(--light-grey);
@@ -68,19 +84,27 @@ header {
   align-items: center;
   line-height: 1;
 }
+
 .logo--square img {
   width: 90px;
   height: 90px;
   margin-top: -1.3em;
 }
+
 .logo--square > div {
   margin-top: -1.7em;
 }
-/* utility classes */
-.highlight {
-  color: var(--tertiary-color);
-}
 
+      /* Utility classes */
+      .highlight {
+        color: var(--tertiary-color);
+      }
+
+      .divider {
+        border-bottom: 2px solid var(--primary-color);
+      }
+
+/* Button styling */
 button {
   padding: 0.5em 2em;
   background-color: var(--secondary-color);
@@ -91,72 +115,79 @@ button {
   font-size: var(--large-font);
   cursor: pointer;
 }
-.divider {
-  border-bottom: 2px solid var(--primary-color);
-}
-/* End utility classes */
 
-/* Cart icon styles */
-.cart {
-  padding-top: 1em;
-  position: relative;
-  text-align: center;
-}
+      /* Cart icon styles */
+      .cart {
+        padding-top: 1em;
+        position: relative;
+        text-align: center;
+      }
 
-.cart svg {
-  width: 25px;
-}
-.cart:hover svg {
-  fill: gray;
-}
+      .cart svg {
+        width: 25px;
+      }
 
+      .cart:hover svg {
+        fill: gray;
+      }
+
+/* Cart item count bubble */
 .cart-item-count {
-  width:.8rem;
-  height:.8rem;
+  display: none;
   position: absolute;
   top: 10px;
-  right: -10px;
+  right: 10px;
   background-color: var(--primary-color);
   color: white;
   border-radius: 50%;
-  padding: 2px 6px;
+  padding: 4px 8px;
   font-size: 12px;
-  font-weight: bold;
 }
 
-@keyframes cartBounce {
-  0%, 100% { transform: scale(1); }
-  50% { transform: scale(1.2); }
-}
+      /* Cart icon animation */
+      .cart-animation {
+        animation: bounce 0.3s ease;
+      }
 
-.cart-animation {
-  animation: cartBounce 0.5s ease-in-out;
-}
+      @keyframes bounce {
+        0% {
+          transform: scale(1);
+        }
+        50% {
+          transform: scale(1.2);
+        }
+        100% {
+          transform: scale(1);
+        }
+      }
 
-/* End cart icon styles */
-
+/* Mission section */
 .mission {
   padding: 0 0.5em;
-  /* margin: 0.5em 0; */
   line-height: 1.3;
   max-width: 600px;
   margin: auto;
 }
-.products {
-  padding: 0.5em;
-}
-.product-list {
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: space-between;
-  list-style-type: none;
-  padding: 0;
-}
-.product-list a {
-  text-decoration: none;
-  color: var(--dark-grey);
-}
-/* Styles for product lists */
+
+      /* Product list styling */
+      .products {
+        padding: 0.5em;
+      }
+
+      .product-list {
+        display: flex;
+        flex-flow: row wrap;
+        justify-content: space-between;
+        list-style-type: none;
+        padding: 0;
+      }
+
+      .product-list a {
+        text-decoration: none;
+        color: var(--dark-grey);
+      }
+
+/* Individual product card */
 .product-card {
   flex: 1 1 45%;
   margin: 0.25em;
@@ -164,6 +195,7 @@ button {
   border: 1px solid var(--light-grey);
   max-width: 250px;
 }
+
 .product-card h2,
 .product-card h3,
 .product-card p {
@@ -177,24 +209,23 @@ button {
 .card__name {
   font-size: 1em;
 }
-/* End product list card */
 
-.product-detail {
-  padding: 1em;
-  max-width: 500px;
-  margin: auto;
-}
+      /* Product detail section */
+      .product-detail {
+        padding: 1em;
+        max-width: 500px;
+        margin: auto;
+      }
 
-.product-detail img {
-  width: 100%;
-}
+      .product-detail img {
+        width: 100%;
+      }
 
-/* Start cart list card styles */
+/* Cart item card styling */
 .cart-card {
   display: grid;
   grid-template-columns: 25% auto 15%;
   font-size: var(--small-font);
-  /* max-height: 120px; */
   align-items: center;
 }
 
@@ -202,9 +233,11 @@ button {
   grid-row: 1/3;
   grid-column: 1;
 }
+
 .cart-card p {
   margin-top: 0;
 }
+
 .cart-card .card__name {
   grid-row: 1;
   grid-column: 2;
@@ -225,27 +258,13 @@ button {
   grid-column: 3;
 }
 
-@media screen and (min-width: 500px) {
-  body {
-    max-width: 1080px;
-    margin: auto;
-  }
-  .mission {
-    font-size: 1.2em;
-  }
-  .cart-card {
-    font-size: inherit;
-    grid-template-columns: 150px auto 15%;
-  }
-}
+      /* Footer styling */
+      footer {
+        font-size: var(--small-font);
+        padding: 1em;
+      }
 
-footer {
-  font-size: var(--small-font);
-  padding: 1em;
-}
-
-/* alert css */
-
+/* Alert modal styling */
 .alert-modal-overlay {
   position: fixed;
   top: 0;
@@ -278,7 +297,7 @@ footer {
 }
 
 .close-button:hover {
-  color: #a4b8c4
+  color: #a4b8c4;
 }
 
 .alert-message {
@@ -286,5 +305,21 @@ footer {
   padding: 10px;
   margin: 10px 0;
   font-size: 18px;
-  
 }
+
+      /* Responsive styles */
+      @media screen and (min-width: 500px) {
+        body {
+          max-width: 1080px;
+          margin: auto;
+        }
+        
+        .mission {
+          font-size: 1.2em;
+        }
+        
+        .cart-card {
+          font-size: inherit;
+          grid-template-columns: 150px auto 15%;
+        }
+      }
